# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/account/getUserInfo": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/account/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登陆",
                "parameters": [
                    {
                        "description": "登录凭证",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/account/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册凭证",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/did/create_identity": {
            "post": {
                "description": "通过传入软件包名称后台自动生成一个did标识符，并生成与其对应的did文档",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "did身份标识管理"
                ],
                "parameters": [
                    {
                        "description": "软件名称",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/did/query_identity": {
            "get": {
                "description": "通过didID来查询软件标识，传入指定的didID可以查询到该指定的文档，想要获取链上全部标识时，参数填“all”会返回所有标识。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "did身份标识管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询单个标识时传入该文档指定的didID，想要获取全部标识时，就直接传入参数“all”。 ",
                        "name": "did",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/did/remove_identity": {
            "delete": {
                "description": "通过表单数据来删除标识信息，注意是表单数据哦！！！",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "did身份标识管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "didID",
                        "name": "didID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/did/update_identity": {
            "put": {
                "description": "通过传入结构体body数据来更新标识信息，注意这个结构体仅仅只是Document这一部分哦，这只是整个标识数据模型中的一部分，切记不要传入整个标识！！！",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "did身份标识管理"
                ],
                "parameters": [
                    {
                        "description": "新的did文档",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DidDocument"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/requestBase.ResponseBodyData"
                        }
                    }
                }
            }
        },
        "/api/v1/pkg/getDetail": {
            "post": {
                "description": "获取软件包相关信息(生成的sbom,漏洞分析报告下载路径)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件包管理"
                ],
                "summary": "获取软件包相关信息(生成的sbom,漏洞分析报告下载路径)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkgUploadController.GetDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "结果",
                        "schema": {
                            "$ref": "#/definitions/pkgUploadController.GetDetailResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/pkg/query": {
            "post": {
                "description": "查询用户上传软件包记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件包管理"
                ],
                "summary": "查询软件包列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkgUploadController.QueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "软件包列表",
                        "schema": {
                            "$ref": "#/definitions/model.PkgRecord"
                        }
                    }
                }
            }
        },
        "/api/v1/pkg/upload": {
            "post": {
                "description": "上传软件包压缩包到服务器并返回访问地址,支持.zip和.tar.gz格式压缩包,格式采用multipart/form-data,字段为file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件包管理"
                ],
                "summary": "上传软件包",
                "parameters": [
                    {
                        "type": "file",
                        "description": "选择要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "$ref": "#/definitions/model.PkgRecord"
                        }
                    }
                }
            }
        },
        "/api/v1/sbom/generate": {
            "post": {
                "description": "根据已上传的软件包生成SBOM文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SBOM管理"
                ],
                "summary": "生成SBOM接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "生成参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sbomController.GenerateSBOMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SBOM清单信息",
                        "schema": {
                            "$ref": "#/definitions/model.SBOMReport"
                        }
                    }
                }
            }
        },
        "/api/v1/vulnerability/scan": {
            "post": {
                "description": "根据已经生成的sbom生成漏洞分析报告",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "漏洞分析"
                ],
                "summary": "生成分析漏洞接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分析参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vulnerabilityScanningController.ScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SBOM清单信息",
                        "schema": {
                            "$ref": "#/definitions/model.ScanReport"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Authentication": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "sign": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "model.DidDocument": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "authentication": {
                    "$ref": "#/definitions/model.Authentication"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "service": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Service"
                    }
                },
                "updated": {
                    "type": "string"
                },
                "verificationMethod": {
                    "$ref": "#/definitions/model.VerificationMethod"
                }
            }
        },
        "model.PkgRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "didid": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "pkgFilename": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "model.SBOMReport": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "didid": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "model.ScanReport": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "didid": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "model.Service": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "serviceEndpoint": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.VerificationMethod": {
            "type": "object",
            "properties": {
                "controller": {
                    "type": "string"
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "pkgUploadController.GetDetailRequest": {
            "type": "object",
            "properties": {
                "didid": {
                    "type": "string"
                }
            }
        },
        "pkgUploadController.GetDetailResponse": {
            "type": "object",
            "properties": {
                "sbomFileUrl": {
                    "type": "string"
                },
                "scanReportFileUrl": {
                    "type": "string"
                }
            }
        },
        "pkgUploadController.QueryRequest": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "requestBase.ResponseBodyData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回的状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据",
                    "type": "object",
                    "additionalProperties": {}
                },
                "msg": {
                    "description": "返回的具体信息",
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "sbomController.GenerateSBOMRequest": {
            "type": "object",
            "properties": {
                "didid": {
                    "type": "string"
                },
                "format": {
                    "type": "string",
                    "enum": [
                        "spdx-json",
                        "cyclonedx-json",
                        "syft-json"
                    ]
                },
                "pkg_file_url": {
                    "type": "string"
                }
            }
        },
        "vulnerabilityScanningController.ScanRequest": {
            "type": "object",
            "properties": {
                "didid": {
                    "type": "string"
                },
                "sbom_file_url": {
                    "type": "string"
                }
            }
        }
    }
}